Remove file selector, rename to folder selector
